+++ projects/start-sdk/lib/StartSdk.ts
@@ -77,7 +77,7 @@ import { createVolumes } from './util/Volume'
 import { getDataVersion, setDataVersion } from './version'
 
 /** The minimum StartOS version required by this SDK release */
-export const OSVersion = testTypeVersion('0.4.0')
+export const OSVersion = testTypeVersion('0.4.0.2')
 
 // prettier-ignore
 type AnyNeverCond<T extends any[], Then, Else> = 
@@ -144,6 +144,8 @@ export class StartSdk<Manifest extends T.SDKManifest> {
       | 'clearServiceInterfaces'
       | 'bind'
       | 'bindRange'
+      | 'retireHost'
+      | 'retireBinding'
       | 'getHostInfo'
     type MainUsedEffects = 'setMainStatus'
     type CallbackEffects =